Start new branch from 0 to remake the game logic
True modular design should be capable of being
reused, for example, my function to make a grid
should be usable to make a chess board, that's not
how I was implementing functions and object
factories.

If I am gonna learn modular design, I have to
learn it well, this new branch purpose is to do
just that, make truly reusable modules.