End version 2 of tic-tac-toe
Although more reusable modules have been made,
which is already a big improvement respect the
version 1, the design of the whole system is a
mess.

The reason to abandoning this version before
completing it, is the believe that I can make a
more clean system, as I think that the problem
that ended up getting the system messy was the
lack of pre-planning before coding the whole thing.

I hope that with a little bit of pseudocode and
design work, I could make a much better modular
board game system.